cryptography/requirements.txt
cyhunspell >= 2.0.2